feat: add DTOs, mappers, services e controllers prontas

+++ Exercícios/10-30/biblioteca.sptech.school/src/main/java/biblioteca/biblioteca/sptech/school/entity/Livro.java
@@ -10,10 +10,9 @@ import java.time.LocalDate;
 @Getter
 @Setter
 public class Livro {
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     private String nome;
